requirements.txt
aiohttp==3.8.4 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
alembic==1.10.4 ; python_version >= "3.10" and python_version < "4.0"
amqp==5.1.1 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
asyncpg[sqlalchemy]==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.10" and python_version < "4.0"
billiard==3.6.4.0 ; python_version >= "3.10" and python_version < "4.0"
celery==5.2.7 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
click-didyoumean==0.3.0 ; python_version >= "3.10" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
click-repl==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cryptography==40.0.2 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.10" and python_version < "4.0"
fastapi-cache2[redis]==0.2.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.95.1 ; python_version >= "3.10" and python_version < "4.0"
fastapi[all]==0.95.1 ; python_version >= "3.10" and python_version < "4.0"
flower==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and platform_machine == "aarch64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "ppc64le" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "x86_64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "amd64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "AMD64" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "win32" and python_version < "4.0" or python_version >= "3.10" and platform_machine == "WIN32" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
humanize==4.6.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
kombu==5.2.4 ; python_version >= "3.10" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.8.11 ; python_version >= "3.10" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
pendulum==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
prometheus-client==0.16.0 ; python_version >= "3.10" and python_version < "4.0"
prompt-toolkit==3.0.38 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pytzdata==2020.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
redis==4.5.4 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==67.7.2 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqladmin==0.10.3 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.12 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.26.1 ; python_version >= "3.10" and python_version < "4.0"
starlette[full]==0.26.1 ; python_version >= "3.10" and python_version < "4.0"
tornado==6.3.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
ujson==5.7.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn[standard]==0.22.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
vine==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.2 ; python_version >= "3.10" and python_version < "4.0"
wtforms==3.0.1 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.11" and python_version < "4.0"
